\name{example}
\alias{example}
\title{example dataset}
\description{

A simulated data set composed by 500 covariables and 150 
values of response variable.
}

\usage{
  data(example)
}

\format{

 \code{markers} is the incidence matrix.
 \code{yield} is the vector of response variable.

}


\keyword{datasets}
