\name{plot.BayesH}
\alias{plot.BayesH}
\title{Plots observed versus predicted values}
\description{
    Plots observed versus predicted values for objects of class BayesH.
}
\usage{
\method{plot}{BayesH}(x, ...)
}

\arguments{
  \item{x}{An object of class \code{BayesH}.}
  \item{...}{Further arguments passed to or from other methods.}
}

\author{
  Renato Rodrigues Silva
}

\details{ 
    It is an adaptation of code written by Perez and De Los Campos (2014). 
}

\references{ Gustavo de los Campos, Hugo Naya, Daniel Gianola, Jose Crossa, Andre's Legarra,
             Eduardo Manfredi, Kent Weigel and Jose Miguel Cotes. Predicting Quantitative Traits 
             With Regression Models for Dense Molecular Markers and Pedigree, 2009. Genetics. 
             doi: 10.1534/genetics.109.101501
             
           }

\keyword{plot}
